Clarify what connection pools do
Add separate section on Configuring SSL
Remove method signature from Configuration Options; this detail is available in intellisense on the client and doesn't need repeating in docs.
Split out modifying the default serializer into a separate section so that we can create an anchor list to the sections under the Serialization header.